Lab2-C++ and OOP - Continued
I continued the progress I made in the lab for this lab sheet. I am now about to start Task-344 which seems to be where Nick got to in the lecture notes, the complex creating virtual hardware. I will continue this when I return to the flat from going home tomorrow :).